renamed islami scaffold & cleaned SuraList.dart
