+++ lib/Tasbe7.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:islami/utility/islami_scaffold.dart';
+import 'package:islami/utility/IslamiScaffold.dart';
 import 'package:islami/BottomNavBar.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 class Tasbe7 extends StatefulWidget {
