+++ lib/quran/QuranScreen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:islami/quran/SuraList.dart';
 import 'package:islami/BottomNavBar.dart';
-import 'package:islami/utility/islami_scaffold.dart';
+import 'package:islami/utility/IslamiScaffold.dart';
 class QuranScreen extends StatefulWidget {
   QuranScreen({Key? key});
   static const routeName = "/quran_list";
