+++ lib/quran/SuraList.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:islami/quran/SuraLink.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../MyTheme.dart';
@@ -93,101 +92,3 @@ class SuraList extends StatelessWidget {
     );
   }
 }
-
-// class _SuraListState extends State<SuraList> {
-//
-//   late Future<List<Widget>> futureSuraList;
-//
-//   Future<List<Widget>> buildSuraList() async {
-//
-//
-//     final List<Widget> suraList = [
-//       Container(
-//         alignment: Alignment.center,
-//         decoration: BoxDecoration(
-//             border: Border(
-//               top: SuraList.borderSide,
-//               bottom: SuraList.borderSide,
-//               right: SuraList.borderSide,
-//             )
-//         ),
-//         child: Text('surah name'),
-//       ),
-//       Container(
-//           alignment: Alignment.center,
-//           decoration: BoxDecoration(
-//               border: Border(
-//                 top: SuraList.borderSide,
-//                 bottom: SuraList.borderSide,
-//               )
-//           ),
-//           child: Text('number of verses')
-//       )
-//     ];
-//
-//     String data = await getFileData("assets/sura_names.txt");
-//     List<String> suras = data.split(",");
-//
-//     data = await getFileData("assets/sura_size.txt");
-//     List<String> suraSize = data.split(" ");
-//
-//
-//     suras.map((sura) {
-//       return sura.replaceAll('"', '');
-//     }).toList().asMap().forEach((index, sura) {
-//       suraList.add(
-//           SuraLink(
-//             suraNumber: index + 1,
-//             decoration: BoxDecoration(
-//                 border: Border(
-//                   right: SuraList.borderSide,
-//                 )
-//             ),
-//             text: suraSize[index], suraName: sura,
-//           )
-//       );
-//
-//       suraList.add(
-//           SuraLink(
-//             suraNumber: index + 1,
-//             text: sura, suraName: sura,
-//           )
-//       );
-//     });
-//
-//     return suraList;
-//   }
-//
-//   @override
-//   void initState() {
-//     super.initState();
-//     futureSuraList = buildSuraList();
-//   }
-//
-//   @override
-//   Widget build(BuildContext context) {
-//     return FutureBuilder<List<Widget>>(
-//         future: futureSuraList,
-//         builder: (context, snapshot) {
-//           if (snapshot.hasData){
-//             return Expanded(
-//               flex: 3,
-//               child: Directionality(
-//                 textDirection: TextDirection.ltr,
-//                 child: GridView.count(
-//                   crossAxisCount: 2,
-//                   childAspectRatio: 4/1,
-//                   children: snapshot.data!.toList(),
-//                 ),
-//               ),
-//             );
-//           }
-//           else if (snapshot.hasError){
-//             return Text("${snapshot.error}");
-//           }
-//
-//           return const CircularProgressIndicator();
-//         }
-//     );
-//   }
-// }
