+++ lib/sura_content/SuraContent.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:islami/utility/get_file_data.dart';
-import 'package:islami/utility/islami_scaffold.dart';
+import 'package:islami/utility/IslamiScaffold.dart';
 
 class SuraContent extends StatefulWidget {
   const SuraContent({Key? key, required this.args}) : super(key: key);
